Adapt the functions being used from require("daemon").
This module wasn't actually making full use of the daemon node, so it
didn't make sense to pull all its baggage along.

Also, added a bit of prep code for an eventual future where stderr is
left open and attached to a "last ditch" error file. This is done
because Node doesn't provide a way to redirect its "death throes"
messages from the underlying stderr stream. So, right now, we never get
to see any process exit exceptions when running as a real daemon, which
sucks.